quote

+++ hello_world/101-quote.c
@@ -11,10 +11,10 @@
 
 int main ()
 {
-        int i;
-        char str[] = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
+	long unsigned int i;
+	char str[] = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
 
-        for (i = 0; i <= strlen(str); i++)
-                putchar(str[i]);
-        return (1);
+	for (i = 0; i <= strlen(str); i++)
+		putchar(str[i]);
+	return (1);
 }
